Moving echonest module to thirdparty
