+++ tests/test_echonest.py
@@ -14,11 +14,11 @@ import httplib
 # Third Pary Libs
 import mock
 import pytest
-import requests
 from simplejson import JSONDecodeError
 
 # First Party Libs
-from fm.echonest import EchoNestError, get_artist_genres
+import requests
+from fm.thirdparty.echonest import EchoNestError, get_artist_genres
 from tests import ECHONEST_ARTIST_GENRES
 
 
@@ -33,7 +33,7 @@ class TestGetArtistGenres(object):
         self.requests.ConnectionError = requests.ConnectionError
 
         patch = mock.patch(
-            'fm.echonest.requests',
+            'fm.thirdparty.echonest.requests',
             new_callable=mock.PropertyMock(return_value=self.requests))
 
         patch.start()
